+++ Sources/MapboxSearch/PublicAPI/Use Cases/Category/Models/Category+Options.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-extension Discover {
+extension Category {
     public struct Options {
         /// Maximum number of results to return.
         /// The maximum allowed value for SBS APIs is 100 results.