+++ Sources/MapboxSearchUI/HotCategoryButton.swift
@@ -1,5 +1,14 @@
 import UIKit
 
+extension SearchCategory {
+    static let empty = SearchCategory(
+        canonicalId: "",
+        name: "?",
+        legacyName: "unknown",
+        iconName: Maki.marker
+    )
+}
+
 class HotCategoryButton: UIControl {
     @IBOutlet private var categoryButton: UIButton!
     @IBOutlet private var textLabel: UILabel!